requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
attrs==24.2.0
audioread==3.0.1
backports.functools-lru-cache==2.0.0
blinker==1.8.2
Brotli==1.0.9
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
cn2an==0.5.22
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
demucs==4.0.1
dora_search==0.1.12
einops==0.8.0
eng_to_ipa==0.0.2
exceptiongroup==1.2.2
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.13.1
fonttools==4.55.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
google-auth==2.36.0
google-auth-oauthlib==1.0.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
indic-transliteration==2.3.37
inflect==7.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
julius==0.2.7
kiwisolver==1.4.7
ko-pron==1.3
lameenc==1.7.0
librosa==0.9.1
llvmlite==0.41.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.5
mdurl==0.1.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
more-itertools==10.5.0
mpmath==1.3.0
narwhals==1.14.2
networkx==3.1
num-thai==0.0.5
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
openai-whisper==20240930
OpenCC==1.1.1
openunmix==1.2.1
orjson==3.10.11
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pip==24.2
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pooch==1.8.2
proces==0.1.7
protobuf==5.28.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pyopenjtalk==0.3.4
pyparsing==3.1.4
pypinyin==0.53.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.4.3
retrying==1.3.4
rich==13.9.4
roman==4.2
rpds-py==0.20.1
rsa==4.9
ruff==0.8.0
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
starlette==0.41.3
streamlit==1.40.1
submitit==1.5.2
sympy==1.13.2
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.12.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.0
transformers==4.46.3
treetable==0.2.5
typeguard==4.4.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.3
uvicorn==0.32.1
watchdog==4.0.2
websockets==12.0
Werkzeug==3.0.6
wheel==0.44.0
win-inet-pton==1.1.0
zhipuai==2.1.5.20230904
zipp==3.20.2
